--- docs/bootstrap.ts
@@ -1,18 +1,0 @@
-import Drash from "../mod.ts";
-
-// Add a global console logger because server logging when needed is cool
-Drash.addMember(
-  "ConsoleLogger",
-  new Drash.Loggers.ConsoleLogger({
-    enabled: true,
-    level: "debug",
-    tag_string: "{date} | {level} |",
-    tag_string_fns: {
-      date: function() {
-        return new Date().toISOString().replace("T", " ");
-      }
-    }
-  })
-);
-
-export default Drash;
